Move machine translation feature here from wishlist-intake repo
Move the article-space machine translation banner feature from
the wishlist-intake repo into this extension, and add some of
the initialization that's required.

The banner is enabled via the behaviour
switch __MACHINETRANSLATION__ being added to a page.

Add a $wgCommunityRequestsEnable feature flag behind which to hide
all functionality of this extension (for now).

Bug: T366194
Change-Id: Ia5c9934bb2319eba4f4d4f2e56298452a273a663