Register no-op parser tags and add database tables
This lays some of the groundwork for what is to follow. The schema is
intentionally vague at this time and subject to further changes. Efforts
are made to prevent data redundancy to MW tables. 'Created at' and
'author' however are stored as they may be overwritten (i.e. for a wish
imported from an older survey). Wikitext-based fields are not stored as
CirrusSearch should do a better job at that, and we want to keep in line
with wiki pages as the primary storage medium; We use our db more for
sorting and filtering purposes on index pages.

The parser tags are added as placeholders and are no-ops, but still
respect $wgCommunityRequestsEnable.

Bug: T366194
Change-Id: Ie2797bbff19293ee5bae272e7269e4b95c2b6277