Add switch button and lang name to machine translation banner
Fix the module dependency name and update the banner message to
include the current interface language's name.

Change the toggle button to be a toggle switch, and reorganise
the messages so that the switch has a label.

This keeps CommunityRequests in sync with the gadget, per
https://gitlab.wikimedia.org/repos/commtech/wishlist-intake/-/merge_requests/268
(which also uses the new messages from here).

Bug: T371098
Change-Id: Ie1af8ef6bd319dc3d94630c0cac7664e71c11268