Add missing 'unknown' wish type label
In the form, the 'unknown' wish type is called "I'm not sure…",
but elsewhere it needs to just be 'Unknown'. The other three type
labels are fine for use (although we do have a discrepency between
'bug report' and 'bug fix' in Template:Community_Wishlist/Type,
which I'm sorting out shortly).

Bug: T370230
Change-Id: I3a2d5629da1078fef85364d8b7b000669e6031c1