styles: Harden design token usage in calculations
As design token values could be numeric values with
units, but also upcoming CSS custom properties,
it's useful to rely on CSS `calc()` function to
ensure that the values are correctly calculated.
This will also bring us closer to remove Less parser
in future and rely on CSS standard only.

Bug: T389234
Change-Id: I432a041b77d0cfb2a0ff82f523c5c05257c0ad3b