WishlistIntake: add config settings for projects, wish type, and status
We want to store using enum values set by config settings, both for
sysadmin flexibility and to minimize the storage footprint.

The statuses need a default value, so an associative array is used for
this setting so they can be used like constants. The older constants
that live in the JavaScript Wish class are removed.

Use a ResourceLoader factory method to dynamically add configurable
messages.

Bug: T387957
Change-Id: Iecf9728edf59ffaa3f9e9f5edfe91855b34ec208