WishStore: add storage layer and value class for wishes
The Wish value class takes a Title as its main argument. This is
expected to be either the wish page itself or a translation subpage.

WishStore::getWish() and ::getWishes() both will attempt to fetch wishes
in the user's language or its fallback chain. If no suitable
translations are found, the base language is used.

The many-to-many relationships - projects and Phab tasks - are bulk
queried separately from the main query to keep the row count within
reason.

Move Wish-related classes to its own namespace, and make IdGenerator a
namespace. Also use the namespace HookHandler instead of Hooks since
they are in fact handlers and not hooks.

Bug: T393827
Change-Id: I732b059e9e8b9a74c5a37bd832e70c647bb2d59e