Render wishes server-side instead of through the template
WishHookHandler::renderWish() is now fully responsible for the display
of wishes. ::renderWishInternal() holds the view logic. It and its
helper methods are necessary until we have Codex PHP to simplify HTML
generation (T379662). The wish template is still needed however, so we
include it as an XML file and install it in SchemaHooks. This hard-coded
to use Template:Community_Wishlist/Wish. Sysadmins will be instructed to
move the page and update $wgCommunityRequestsWishTemplate accordingly if
they wish to use a different title.

This patch adds more situations to put the wish page in the error
tracking category, and also adds a general tracking category for all
wishes. The category is only added on wish pages and not on any page
where the parser tag is transcluded, such as the wish template.

Add new 'ext.communityrequests.styles' style-only module to avoid FOUCs
and give minimal styling to the wish page. This same module will later
be used for essential styling of other pages, too.

Other changes:
* WishHookHandler / init.js: Fix to only show post-edit banner after
  using Special:WishlistIntake.
* SpecialWishlistIntake: Add loading state and fallback for no-JS users,
  making use of previously unused messages.
* WishlistConfig: add getters for labels.
* Get the skipped Selenium test to pass, finally.

Bug: T394355
Change-Id: I3d9357677b0135e3b371d12425fd40f0f0c82da6
Depends-On: I8ccf6d7d903f477cafe21510a7c1e156695343ca