SpecialWishlistIntake: add hidden fields for missing wish data
This fixes a bug where the 'status', 'created' and 'proposer' fields
would get wiped out and re-set when using Special:WishlistIntake. For
'status', this would only happen for non-staff. To fix this, add hidden
input fields so that they get submitted and passed to the Wish
constructor. At a later time, we may expose 'created' and 'proposer' as
staff-only fields (The only use case is for importing wishes from an
older survey, so wikitext editing will suffice for now).

Other changes:
* Make use of WishlistConfig methods instead of using the main Config
  directly.
* Don't use self-closing tag for raw inputs, and eslint is happy

Bug: T394355
Change-Id: I9097fd07859566d00217917bd725c9d7afd8b99f