Display 'All projects' instead of listing each one individually
In the database, we want to store each project individually to make for
easier querying. However when viewing a wish, we want it to say "All
projects" instead of listing each one individually. Similarly, the
wikitext value for this should remain as 'all', for readability but also
for backwards compatibility with wishes created via the legacy gadget.

Also remove the 'revision' table from the WishStore::save() query, which
was left by mistake. The storage layer still tries to fetch the proposer
and created timestamp from the wish record. This is left as an ultimate
safeguard; during normal usage, we should never have a wish edited
without a created timestamp and a valid proposer. In a future patch, we
will make use of the EditFilterMergedContent hook to enforce this for
wikitext editors, akin to Extension:AbuseFilter.

Bug: T394355
Change-Id: I226e64dd12f664005f1873af9401bfe9c2a1b5e0