Improve detection of if a page is a wish or focus area
Before we were merely checking if the title began with the prefix. This
takes it a step further and looks for a numeric ID to follow, or an ID
followed by a slash and valid-looking language code. It is still not
foolproof, but should be good enough given we are avoiding an actual
comparison with the full list of language codes and want WishlistConfig
to remain in a static context.

Change-Id: I9a980e183f8af7f517368dc7c14b3bfdbc2929f5