Abstract wishes/FAs, change page language on creation, fix deletion
NOTE: This patch requires you delete Template:Community_Wishlist/Wish
and re-run the update maintenance script.

Wish and FocusArea, and their stores, all work very similarly. This
patch abstracts common properties and logic from the value classes,
stores, and APIs into the parent abstract classes.

AbstractWishlistStore imposes getters for various db fields so that we
can construct the queries in one place, ensuring consistent behaviour
between wishes and focus areas.

Change page language when creating wishes/FAs in lang other than
default. To make this easier, add 'baselang' as a required field in the
templates. This way we don't need to mess with the Request object to get
the user's language. SpecialWishlistIntake.vue instead is responsible
for setting this field (and later, SpecialEditFocusArea.vue).

In the hook handlers, use the RevisionDataUpdatesHook to determine if
we're creating a new page, and directly set the page.page_lang field if
the 'baselang' differs from the page language. This is only done if
$wgPageLanguageUseDB is set. It is set everywhere Translate is enabled
on our cluster, but this isn't necessarily the case for non-WMF wikis.

Add CommunityRequestsHooks::getCanonicalWishlistPage() which ensures we
use the correct page ID in storage.

For now, creating a wish in a language different than the site language
is not supported when Extension:Translate or PageLanguageUseDB are not
enabled. We need to also prevent changing of the base language after a
page has been marked for translation. This is saved for T398371.

Fix tests to properly use Extension:Translate and add translation
subpages.

This patch also fixes a bug in WishHookHandler where an entire wish and
associated data would get deleted when only a translation subpage was
deleted. The deletion logic has been moved to CommunityRequestsHooks and
now works for both wishes and focus areas.

Other changes:
* Add tracking category for focus areas.
* Add missing config settings for focus areas.
* Add missing 'owners' and 'volunteers' fields to FocusArea (these are
  not stored in the db).
* Add TAG_ATTR_* constants to FocusArea and make use of them in
  FocusAreaHandler.
* Whenever we have lines of code representing fields or properties of a
  wishlist entity, use the same order as the form.
* Keep the names of properties consistent, such as 'lang' instead of
  'language', and 'baseLang' instead of 'baseLanguage'. "Language" can
  imply a Language object, while the shortened name refers to language
  codes.
* ApiFocusAreaEdit: add edit API to parity ApiWishEdit
* Add ApiWishlistEntityBase to share logic between the two edit APIs.
* Make the necessary methods abstract in AbstractWishlistEntity, and
  implement them in FocusArea.

Bug: T396358
Bug: T396361
Bug: T387961
Change-Id: I50a9798806d5963706f6d972837bea0e86a4d1ae
Depends-On: Icb194e92ae65935cc20ed6ec57b2d615305ce53d