ApiQueryWishes: add list API to get a sorted list of wishes
Uses 'communityrequests-wishes' to ensure a stable namespace, with 'crw'
as the prefix.

AbstractWishlistStore::getAll() now sorts by 'title', 'votecount', and
either 'created' or 'updated'. Sub-sorting is chosen based on what we
deem to be the most useful.

Add 'voteCount' and 'updated' to Wish::toArray() for the API's purpose.
The same should eventually be done for FocusArea, but it is not
important for now as we will not have focus area pagination for quite a
while.

WishStore: fix errors that happened when passing in empty arrays to
::getProjectsForWishes() and ::getPhabTasksForWishes().

Make sure we're using strict mode in all PHP classes, and with return
types on all applicable methods.

Bug: T387961
Change-Id: Iec0d1ebb4105e34618e6bc58795a8b26ffd3e32a