Add WishIndexTemplateRenderer and WishIndexTable; fix issue with LIMIT
This implements {{#CommunityRequests: wish-index}}. For now, it just
outputs the DIV for the WishIndexTable.vue to mount to.

Fix a bug with AbstractWishlistStore::getAll() where the LIMIT was one
less than it should be (not allowing for the baselang to be included in
the results). Add test case to reinforce. Also fix sorting issues.

Similar to I2cccf6a2f7, we fetch the focus area piecemeal in
ApiQueryWishes when it could be included in the initial query. TODOs
left accordingly.

A 'crwcount' parameter is added to the API to include the total count of
wishes.

A Codex bug was discovered where the 'first' and 'last' events of the
Table component are not fired. Resultantly, pagination support in Table
might get reworked, so I am not writing Jest tests yet. As an interim
solution, we check the offset and total row count to infer if we're on
the first or last page, and set the direction accordingly.

Bug: T387962
Change-Id: I4b94d9f36bca92dcdbbc4a5a1ffe1b59106d4ca3