Count votes on focus areas and wishes
Subpages of a wish or focus area that match the config for
`pageVotesSuffix` will be considered voting pages and each call to the
{{#CommunityRequests: vote}} parser function with the vote argument will
be counted as vote towards the wish or focus area

This patch does not include adding the right wikitext to the votes pages
via the current UI, and it does not render the vote markup either.
It also does not check for the status of the wish or focus area before
updating the votes since we rely on abuse filter and UI checks for this.

Bug: T388220
Change-Id: I3fbf3f81fc5099a233dd9fce380d43bc4f84303e