CommunityRequestsHooks: show 'Edit with form' tab on wish and FA pages
A new permission 'manually-edit-wishlist' is introduced. If the user
does not have the permission, we hide the native 'Edit' tabs. 've-edit'
for VisualEditor is hard-coded for now, as the extension still has a
hard dependency on VE.

In a separate patch, we will prevent editing of wishes and FAs through
the edit API and other means.

Bug: T400187
Change-Id: I427408c9a025d96cafb7048ac06721c367bac3a4