CommunityRequestHooks: keep track of language in extension data
If a wish is created in a language other than the site language,
the page language in the database is updated during the
RevisionDataUpdates hook handler. However, the DB operation is made in a
deferred update, so any hooks that follow may get the wrong language.
Work around this by setting an initial language as dictated by the
parser, and update after we change the language in the DB.

Use Parser::getTargetLanguage() instead of ::getContentLanguage()

Bug: T400410
Change-Id: I6bc1a4cb904450d503c97d8f5b25ca4c3a157eb3