Set the display title on wish and focus area pages
Move the status chip to be an indicator

Change the redundant <h2> that had the title to be "Description", and
remove the <h3> for "Description".

Vector 2022 forces indicators to be beneath the title, so we have
CSS overrides to make the indicator inline with the subpage subtitle.
Wide content in #contentSub will wrap instead of overlapping with the
status chip.

Bug: T403307
Co-Authored-By: MusikAnimal <musikanimal@gmail.com>
Change-Id: I9e1f0e3020fa99efd71388ce991ca51d2b942fcc