Show message with the updated vote count on wish and focus area pages
Before we were just showing dummy data. We want presentation logic to
live in the renderers (AbstractTemplateRenderer::getVotingSection() in
this case), but since the /Votes subpage is transcluded at the bottom,
we don't have the vote count until after everything is rendered.

Add a strip marker where we want the message to go, and make use of the
ParserAfterTidy hook to set the message with the vote count stored in
the extension data.

Remove the voting button and leave only the container, placing
responsibility of user rights checks etc. solely on the Vue app.

Add a #Voting ID to the section heading.

Bug: T400972
Change-Id: Id11d6a5fad9eb00d33bb1a2be4f027f372b94767