WishlistConfig: use LanguageNameUtils service to verify lang codes
Discovered while investigating T401256.
LanguageCode::isWellFormedLanguageTag() is insufficient.
LanguageNameUtils does not make DB queries but may read the filesystem
when looking for language files. This is only done for language codes
not already in MediaWiki\Language\Data\Names, which should account for
the vast majority, so the performance impact should be minimal.

Remove FIXME in AbstractWishlistStore about file sorting. It's expected.

Bug: T401265
Change-Id: Icda59140baa9d4fe2cbee755d89c0de3c2ec49bd