ApiQueryWishes: link to translation subpage when applicable
If Extension:Translate is installed, the API response should give a
'crwtitle' for the translation subpage if it is available for the
requested language.

Add more test coverage, and fix the namespace in a few files. For now,
CommunityRequestsIntegrationTestCase extends ApiTestCase even tough it
is used for many tests that aren't related to the API. We need easy
access to $this->insertWish(). In a future commit these helps should
perhaps be moved to a trait.

Bug: T401256
Change-Id: Ia9202c85cc013d670eda5c36bccf74ed5a8908f5