WishIndexTable: hide 'last page' button and rm related logic
It looks like Codex won't properly support directionality of
server-generated results for a bit (T401027). Instead we will hide the
"last page" button and enforce on a 'forward' directionality for now.

Additionally, Codex seems to do the math incorrectly when changing the
pagination size and the `serverPagination` option is set. We will
instead allow the provided `limit` (default 10) to be the pagination
size, and hide the dropdown entirely.

Bug: T387962
Change-Id: I0165592de49656c696af96b10dafd241b92d7fc7