AbstractTemplateRenderer: add a 'style' prop to configuration
It's reasonable to assume every CommunityRequests installation will have
a "done" status, but if it is non-English, they will likely use the
translation for "done", so we shouldn't hard-code against it.

Add a "style" prop for entries of $wgCommunityRequestsStatuses,
defaulting to "notice".

Have the wish-index and focus-area index pages use this as well.
The wish-index page is a Vue app, and CdxInfoChip will style it with an
icon, which I assume should be in the PHP implementations, but we can
save that for when we have Codex PHP.

Bug: T400190
Change-Id: I7dadd6e287d12078c50a199e3baa22e15db88039