Supply original wikitext to forms, rework stores and wikitext fetching
Move the caller for fetching raw wikitext from the API to the store. A
new argument for AbstractWishlistStore::get() and ::getAll() indicates
which page to fetch and parse. This takes into account transformations
Extension:Translate may make, while also not depending on Translate
functionality (as opposed to stripping <translate> tags).

For performance, we only fetch such fields when necessary, enforced by
the preexisting AbstractWishlistStore::getWikitextFields().

AbstractWishlistStore::getEntitiesFromLangFallbacks() is renamed to
::getEntitiesFromDbResult(), and our internal method now calls the
abstract method instead of the reverse. A new abstract
::getExtTranslateFields() is added to account for all fields where
<translate> tags may be present. Wikitext fields are dynamically added
as properties to the DB result object, with "virtual" column names to
ease use by subclasses.

Bug: T401264
Change-Id: I90d3ae59290d52fe7de4c2e14e0641ac6cf37a50