Always use entity constants instead of hard-coding strings
Since the beginning, we've had a few variations of which parameter and
argument names to use. For example, when constructing a Wish, we'd pass
in 'voteCount' or 'baseLang', but in the APIs and other places, we want
the lowercase form 'votecount' and 'baselang'. I believe this to be
partly at fault for the mysterious "blanking of values" issue we've been
facing. So, let's nip it in the bud and use just one variation to rule
them all. All accessors should use PARAM_ constants from the entity
classes. Yes, we lose our camelCasing but so be it. The API will only
accept lowercase parameter names, so that is what we use.

This patch is in preparation of I90d3ae5929

Bug: T401264
Change-Id: I621c8d0aa1efc6c11440a40d9db34e73c45d5def