tests: move shared logic to a triat; run Translate-related jobs
CommunityRequestsIntegrationTestCase and the many characters that make
up its class name are no more, replaced by WishlistTestTriat, which only
requires that classes extend MediaWikiIntegrationTestCase. The trait
consolidates everything needed to properly create entities and to add
translations. PHP's single inheritance makes the trait a better fit than
an abstract class.

Require callers of WishlistTestTrait::inserTestWish() to be more
explicit in what they want while allowing more flexibility. Simply
having a field with <translate> tells it you want it marked for
translation, and now the jobs are actually ran, guaranteeing via
assertions that the language subpage exists.

To edit a translation subpage, we use the same system as the Wish and
FocusArea constructors: you give ->insertTestWish() the root entity
page, specifying it's language with Wish::PARAM_BASE_LANG, and the $lang
argument specifies the translation subpage that will be edited.

After these changes, a bug surfaced: when deciding whether to change the
page language, CommunityRequestsHooks::onRevisionDataUpdates() was
checking if the page was NOT yet setup for translation, when it should
be checking if it is the base entity page (not any translation subpage –
even the base language subpage). This also removes yet another runtime
dependency on Extension:Translate, which is good.

Bug: T401264
Change-Id: I597a0c0d92955d005232614e5b1eafb43d0e166d