ApiWishlistEntityBase: validate wikitext when saving
Validate that the wikitext we parse can be regenerated into exactly the
same wikitext. This is in effect validates there are no stray unescaped
pipes or other artifacts that break parsing.

Introduce AbstractWishlistStore::getArrayParams() for each store to
specify which fields are array-like. A new method normalizeArrayValues()
makes use of that to normalize array-like values into the desired
format.

AbstractWishlistSpecialPage is now responsible for doing this
normalization, both before submitting to the API (when we need
pipe-separated values) and loading the form (true arrays). We also do a
bit of refactoring to consolidate common logic in the abstract class.

Similarly, move common logic from the APIs to ApiWishlistEntityBase.

Bug: T401264
Change-Id: I72b1928975001cb63ecb7047902af1ee5cd78242