MinT: remove magic word; add preference to disable machine translation
The __MACHINETRANSLATION__ magic word was added because originally the
extension would require templates for wishes, focus areas, etc., such
that we'd need something in the wikitext to trigger loading the RL
module. Now that everything is loaded via configuration, there's no need
to store new rows in page_props solely for MinT integration. It'd be
nice to have the magic word for general use, but that's outside the
scope of this extension and can be explored separately.

Introduce a new preference for machine translation in the Wishlist,
shown under the "Internalisation" preferences on the "User profile" tab.

Rename the translation banner init script to init.js for consistency,
and move most of the logic to the Vue app because we need to react to
dynamic content like the wish index page. The Vue app is also reworked
to use the Composition API and ES8 syntax.

Syncing changes from the gadget, we now use localStorage to enable
auto-translation. This may seem redundant to the preference, but
presumably given the gadget was removed as default-on, users have a
desire to hide the translation banner entirely.

Finally, we weren't setting a lang= attribute in table cells, so the
translation banner didn't know it needs to translate on that page. Fix
this by adding a PARAM_BASE_LANG prop to ApiQueryWishes, and use that in
the Vue app to determine if we're (probably) showing text in a different
language.

Other changes:
* Remove no-op styles from the render-blocking styles.less

Bug: T401449
Change-Id: I7e5a12e300a1c10830ab060eb84b43c6d2f964f4