tags: replace 'Projects' feature with tags
This is more or less a drop-in replacement. The only difference on the
backend is 'Projects' had an 'All projects' option, and there is no such
option for tags.

$wgCommunityRequestsTags will house all three kinds of tags. For now
there are only 'navigation' tags.

Use a MultiselectLookup for the form field in Special:WishlistIntake.
This is more or less a direct copy of the Codex example at
https://w.wiki/FBfz.

Bug: T400414
Change-Id: I3b3d21f8f0dc0b3d64e803f2c8144a7efa96baf6