SpecialEditFocusArea: invalidate the index page cache after saving FAs
This ensures the focus area index page will (nearly) immediately show
any new FAs or updates to existing FAs.

This is not necessary for the wish index page since it is currently only
a clientside implementation.

I could not get the unit tests to pass in CI. See PS7 for my final
attempt.

Change-Id: I0ecfe7ccbd1436f6be7ddc983d6af5150005c313