AbstractRenderer: categorize translation subpages
It is convention on Meta to use {{#translation:}} in the category names
so that new categories are produced for each language. This is then used
in the "Search all wishes" inputbox with:

  incategory:"Community Wishlist/Wishes{{#translation:}}"

This way, non-English users get only results in their language.

We want these to be registered as normal categories and not tracking
categories. We need to use the site language to get the correct category
name, then append the parsed "{{#translation:}}" which will resolve to
the language code of the page, or a blank string if its the root entity
page (in which case no additional category is added).

Change-Id: I2c3ba9196f1fa2b5fe700ac13f83f61aed587db1