CommunityRequestsHooks: reload entity data on votes pages
This fixes the elusive "blank title" bug. When saving an edit on a
/Votes subpage, the parser data is incomplete. Some fields like the
proposer and created timestamp are not required to be provided, and
WishStore will reload them. Any other extension data-provided fields
however will be wiped out, so we need to reload the entity in such
cases.

We now throw an exception if the title is missing. I did this first as a
means to debug the 'blank title' problem, and given VoteRendererTest
failed, it was immediately obvious that voting was the culprit. Let's
leave the exception in, anyway; If there's ever a reason it becomes
blank at the time of saving, we want to know about it.

Also update WishStore and FocusAreaStore tests to expect the actual
exception message. In doing so we discover they weren't even testing
what they were supposed to! This is now fixed. Finally, remove
::testSaveWithNoCreationDate() as there shouldn't be an exception in
such cases, given the created timestamp is fetched from the entity page
if not provided.

Bug: T403555
Change-Id: I1b90223b8e6ec38242d3e81268a01610b5fb8152