store: strip translate tags FETCH_WIKITEXT_TRANSLATED when no subpage
The AbstractWishlistStore::FETCH_WIKITEXT_TRANSLATED mode is meant to
return each field without translate tags and related syntax, but if the
page has not yet been marked for translation (or FuzzyBot fails), the
subpage with the parsed output won't be present, meaning we use the base
page, and thus we get translate tags when we didn't want them.

Fix this in the stores by using the optional TranslatablePageParser
service and its ::containsMarkup() and ::cleanupTags() methods.

Remove Phan suppressions for Extension:Translate classes. We already
had Extension:Translate in the Zuul config apparently, we just weren't
telling Phan to include it in our phan/config.php

Bug: T403675
Change-Id: I0815213366b4092dd1be235fdaee31e0f47dfc70