stores: Merge wish and focus area tables, and same for translations
communityrequests_wishes and communityrequests_focus_areas now live
together as communityrequests_entities. A cr_entity_type column
specifies if it's a wish or focus area, while the wish type has moved to
cr_wish_type. There is also cr_focus_area and cr_actor that are
exclusive to wishes that now need to be nullable.

Similarly, communityrequests_wishes_translations and the focus area
equivalent are merged together into communityrequests_translations. This
doesn't need a type column, since the primary key is the page ID and we
never fetch translations without having a page associated with them.

Previously we were storing translations for the short description, but
we've never made use of that. The column is not moved to the new
entities table and instead now lives as a 'wikitext field'.

The static methods for the schema in AbstractWishlistStore made this
work a bit easier. Expand on this and use more of the static methods in
the unlikely event schema changes are needed again.

The actual schema change is done manually so we can use RENAME and to
avoid an excessive number of patch files that would be created had we
used abstract schema changes. All data is preserved during the schema
change.

Bug: T404022
Change-Id: I4db03ba364a953759e372b547f510566abd31bb3