Fix API filtering of wishes when passing an invalid focus area
This moves the entity-specific filtering into the child classes
instead of AbstractWishlistStore, and keeps only the status filter
in that class (as it's common to wishes and focus areas).

It checks that the supplied `FAn` IDs are valid, and if any are not
it throws an API error.

I know we don't have a query API for focus areas yet, but presumably
we will at some point. Anyway, it was weird to be checking the
entity type in the abstract class.

Bug: T404265
Change-Id: I3599ffb43025188889ab215671ab60a0c34028bc