AbstractWishlistStore: query Core's page table separately in prep for x1
Remove the JOIN on the page table and fetch the page_namespace and
page_title in a separate query. These fields are necessary for a Wish or
FocusArea object to be constructed. For example, see
WishStore::getEntitiesFromDbResult().

To make this work, we now have to do a first pass through the result set
from the query on our tables to collect the page IDs, then query to get
the namespace and titles, and then re-add them when we group the results
by page ID and language.

Bug: T404190
Change-Id: I7d8228e9a5c6de0b11a80a65e696828184e45e48