voting: use new action=wishlistedit API and add ability to edit/rm votes
First, move the post-edit JS logic from the intake module to the voting
module. This more easily allows to use the same post-edit logic in both
places, and is also a micro-optimization in that we are loading just one
module now on wish detail pages instead of two.

To accommodate the new post-edit message, we now have new values for the
session variable for the creation of a vote, an update, and removal.
They live in CommunityRequestsHooks now since that's where they are
handled. Removing the session value server-side should also fix T404123.

For the API, fetch the base rev ID at the same time we fetch content.
This fixes an issue where for example, if you opened a wish detail page,
read it for 10 minutes, and tried to vote, you wouldn't be able to if
someone else voted during that time.

Match copy and spacing to more closely match the designs.

Rename voting Vue app filenames to match their component name.

Add more tests for the Vote class.

This patch includes i18n key changes: communityrequests-create-success
is now -wish-create-success and likewise for -edit-success.

Bug: T399500
Depends-On: I54a78446deb19131dc25ce0770ac89973c4022e7
Change-Id: I61ee33cb6ce4f592484a0279e1d0f9798623f5ad