Api: Use try/catch when performing parser validations
Depending on where the stray pipe (or whatever) that breaks parsing is
located, we may not even be able to fetch other paramters of the 'fake'
validation entity such as the base lang (which is the last param). This
in turn can cause a different error.

Let's just simplify it by putting a try/catch around the whole
validation. If anything fails during parser validation, it should be
safe to conclude it's an issue with the wikitext.

Bug: T404108
Follow-Up: Ic60bc425cfbc9ba04367b972f8fae2feba825358
Change-Id: Iecdc8209c1ad73aaab5642b4671f18df89aa633b