Add tooltip to "+N more" tags link
Add a tooltip to the "+N more" tags link in the wishes index
table, and link to the #tags section header on the wish page.

The tooltip uses the listToText method, which adds 'and' between
the last two items, so also quote the tag labels to make it clear
when an 'and' might be part of the tag label.

Bug: T405386
Change-Id: Ia3cd4fed9d6035013ebcbe3e12ecd0888f34c84f