Remove the /Wishes and /Focus_areas from page titles
We are able to infer if a page is a wish or focus area based on the
'W' or 'FA' prefix, so there's no need for /Wishes and /Focus_areas to
be part of the path.

Thankfully this was configurable from the beginning, so the only real
change that was need was in extension.json and the parser tests. The
other tests would still pass using the older paths, but they are updated
anyway to avoid confusion.

NOTE: Reviewers will probably want to run the nukeWishlist maintenance
script *before* checking out and testing this patch.

Bug: T405527
Change-Id: I531f2a208973a6fa2a00952cf3c84d9a5d6f4255