Add wish counts to focus area index cards
Add each focus area's total wish count to the cards on the focus
area index. This adds a new DB query to count these, to avoid
adding more entity-specific logic to the AbstractWishlistStore::getAll()
method.

A few system messages are also documented (and fixed), to make it
easier to find them (along with a couple that are actually relevant
to this patch).

Bug: T405391
Change-Id: Ifebb1cd7c3287c134d44267099587f52fa6d0304