NukeWishlist: make more robust and add 'legacy' option
The previous implementation was clunky and error-prone. It should now
work more reliably.

A new --legacy option is added to alternatively delete pages under the
old prefix instead of the new one. Though, this will only be useful for
a day or two more, but it's here if anyone needs it in the interim.

A --dry-run option is added as well.

Change-Id: I4993be6571013959345bf8854b2f2417ea681b56