WishRenderer: short-circuit and show error if proposer is invalid
If the proposer is invalid, the entity won't be saved. Instead of
erroring out in the storage layer, we should show an error to the user
and put the page in the error tracking category.

This will surface existing wishes with invalid users, such as those who
were renamed since their wish was first created.

Use the Maintenance_script user in the parser tests since this user
exists on all MW installations. Also add pages for user and user talk to
make the parser output easier to read.

Bug: T406194
Change-Id: Ie699d311a7bebd45b9ec8548f88e1474386bac53