SearchHooks: show search results using the entity title
Add a handler for the ShowSearchHitHook that replaces the $link with the
entity title and the page title in parentheses, just like Wikibase does.
We also include the number of supporters of the wish, because we can :)
For focus areas, we additionally include the number of wishes.

The page title in parentheses is given a slightly smaller font. We don't
want to create an entirely new RL module just for this, and it's just
one CSS property, so we use inline styles. Our styles module is too big
to be used for this, with the icons it loads.

Bug: T406423
Change-Id: Id6040acf41136054ba4460c6166b11cb1cb3edcf