AbstractWishlistStore: revise main query to ensure we get translations
What we're trying to do in ::getAll() is fetch all rows for the
requested language along with those in its fallback chain, *OR* where
the base language = requested language (i.e. there are no suitable
translations). Then in post-processing, we group by language following
the fallback chain. The latter part works, but the older query sometimes
would not return results in the requested language even when they
existed.

The query is revised to use a NOT EXISTS operator against a subquery
that guarantees you will get back suitable translations if they exist.
More specifically, we filter out results in the base language if there
are any results for the requested language or its fallbacks.

Bug: T406680
Depends-On: I54a78446deb19131dc25ce0770ac89973c4022e7
Change-Id: Iace14e319a6d4c684a2227208490ab277200af0b