Add error message for invalid statuses
Add error message HTML to the indicator output. This looks bad,
but that's okay because at least it tells the editor what the
problem is and why the page has been added to the error tracking
category.

Add a new AbstractRenderer::getErrorMessage() method, and use
it for the new status error message as well as a few other existing
ones.

Bug: T406188
Change-Id: I9ca03886855250be47c16824e28cedcec2f81403