special pages: simplify logic and ensure we have a valid page title
Before, AbstractWishlistSpecialPage::$pageTitle was always a Title
because a null $entityId will type-cast to a blank string when prefixed
with 'Community Wishlist/W'. In such cases the title is valid, but not
the title we wanted!

This patch fixes that. Simplify the logic, remove the $pageTitle
property entirely and instead pass it around. The logic is now that
$pageTitle is never used unless we are given a valid entity ID.

Change-Id: I1504c347cb0405a68552f74fc785ad825d6faaa8