DeleteOrphanedEntityRows: new maintenance script for DB cleanup
There is a bug where CommunityRequests data is not deleted when an
entity page is deleted. Especially if the ID counter is reset (via the
NukeWishlist maintenance script), you can end up with orphaned rows in
the CommunityRequests tables. This script is added to delete those rows.

A test is added to represent the current state of affairs. After
I01fd96145d, the test should fail or be changed to delete rows in the
page table directly instead of using DeletePage (which runs hooks).

Bug: T406059
Change-Id: I51c25d149b9bbb73d315aa16ef3cc00fce7895d0