tests: make WishlistTestTriat easier to use, convert tests to use it
WishlistTestTrait::insertTestWish() and ::insertTestFocusArea() now
don't require any arguments. If the $title is null, we generate a new
ID, and the language is assumed to be 'en'.

Convert some existing tests to use the helper methods.

extension.json: Fix autoloading of NukeWishlist to only be for the tests
and not all environments.

Change-Id: Idb3b2901065d3d400ba58642c8898398c6ba9d72