WishlistMessageLoader: New class for loading configurable messages
It was weird to have the ResourceLoader messages factory method in
SpecialWishlistIntake when it's used in other places, too, so we move it
to a standalone class. I wanted to have it extend FileModule, but then
we need it to extend CodexModule, which makes the extension.json a bit
confusing.

Remove the hard-coded configurable messages from the wish-index module.

Change-Id: I2e73e4d46b863cd7d993b735fc04ad1a3dcfe54a