VoteStore: use regular expression to split votes rather than a newline
This is just a bandage to the actual problem. We should not need any
regular expressions when parsing content. Instead, our ArgumentExtractor
should support returning multiple instances of the given parser
function. Save that for effort for a rainy day :)

One could argue newlines don't belong in votes at all, and that may be
something we explore later. For now, this patch just ensures things
won't break when newlines are used in vote comments.

Bug: T407193
Change-Id: I1606f28a78f6a792b44ee9675e457f59fb522bc3