Sections should be expanded in mobile
We should fire the `wikipage.content` hook only for the community
wishes DOM element, since it's the only area that was updated when fetching wishes.

When we fire this hook other extensions and skins are listening, like
MobileFrontend, and they are applying their own event handlers depending
on the area that was updated. This may cause unexpected behaviour.

Bug: T407044
Change-Id: I5a7707bb398388ecd4079228091e692f8e871d7c