AbstractRenderer: disable DiscussionTools on voting section
DiscussionTools doesn't work on votes, but that's only by circumstance
(see T406995). Add `.mw-notalk` to the voting section to definitively
disable all DiscussionTools functionality for that section. See
https://w.wiki/FE9H for related docs.

In the future, we may want to allow use of DiscussionTools features
(namely topic subscription) to the "Updates" section (T406286), so we
don't want to use the __NOTALK__ magic word.

Change-Id: I57d4d16f32e8306660f667699535325c43caa153